Improvements to persistence

+++ rules/scala/zinc/ZincPersistence.scala
@@ -8,7 +8,7 @@ trait ZincPersistence {
   def save(): Unit
 }
 
-class FilePersistence(cacheDir: Path, analysisFiles: AnalysisFiles, classDir: Path) extends ZincPersistence {
+class FilePersistence(cacheDir: Path, analysisFiles: AnalysisFiles, jar: Path) extends ZincPersistence {
   private[this] val cacheAnalysisFiles =
     AnalysisFiles(
       apis = cacheDir.resolve("apis.gz"),
@@ -17,15 +17,21 @@ class FilePersistence(cacheDir: Path, analysisFiles: AnalysisFiles, classDir: Pa
       sourceInfos = cacheDir.resolve("infos.gz"),
       stamps = cacheDir.resolve("stamps.gz"),
     )
-  private[this] val cacheClassDir = cacheDir.resolve("classes")
+  private[this] val cacheJar = cacheDir.resolve("classes.jar")
+
+  /**
+   * Existance indicates that files are incomplete.
+   */
+  private[this] val tmpMarker = cacheDir.resolve(".tmp")
+
   def load() = {
-    if (Files.exists(cacheDir)) {
+    if (Files.exists(cacheDir) && Files.notExists(tmpMarker)) {
       Files.copy(cacheAnalysisFiles.apis, analysisFiles.apis)
       Files.copy(cacheAnalysisFiles.miniSetup, analysisFiles.miniSetup)
       Files.copy(cacheAnalysisFiles.relations, analysisFiles.relations)
       Files.copy(cacheAnalysisFiles.sourceInfos, analysisFiles.sourceInfos)
       Files.copy(cacheAnalysisFiles.stamps, analysisFiles.stamps)
-      FileUtil.copy(cacheClassDir, classDir)
+      Files.copy(cacheJar, jar)
     }
   }
   def save() = {
@@ -33,12 +39,14 @@ class FilePersistence(cacheDir: Path, analysisFiles: AnalysisFiles, classDir: Pa
       FileUtil.delete(cacheDir)
     }
     Files.createDirectories(cacheDir)
+    Files.createFile(tmpMarker)
     Files.copy(analysisFiles.apis, cacheAnalysisFiles.apis)
     Files.copy(analysisFiles.miniSetup, cacheAnalysisFiles.miniSetup)
     Files.copy(analysisFiles.relations, cacheAnalysisFiles.relations)
     Files.copy(analysisFiles.sourceInfos, cacheAnalysisFiles.sourceInfos)
     Files.copy(analysisFiles.stamps, cacheAnalysisFiles.stamps)
-    FileUtil.copy(classDir, cacheClassDir)
+    Files.copy(jar, cacheJar)
+    Files.delete(tmpMarker)
   }
 }
